Some little improvements - Increased vision sectors to 7 - Added generation counter - ...

+++ Meat.cpp
@@ -10,7 +10,7 @@ Meat::Meat(Coordinate c, double radius, int quantity) : Resource(c,radius,quanti
 
 int Meat::eat(int quantity)
 {
-    int eated = std::min(quantity,m_quantity);
+    int eated = std::min(quantity,(int)m_quantity);
     m_quantity -= eated;
     return eated;
 }
