+++ Resource.cpp
@@ -1,12 +1,12 @@
 #include "Resource.h"
 
-Resource::Resource(double x, double y, double radius, int quantity, int quantityEvolution) : NonSolid(x, y, radius),
+Resource::Resource(double x, double y, double radius, int quantity, double quantityEvolution) : NonSolid(x, y, radius),
     m_quantity(quantity), m_quantityEvolution(quantityEvolution)
 {
     m_maxQuantity = m_quantity;
 }
 
-Resource::Resource(Coordinate c, double radius, int quantity, int quantityEvolution) : NonSolid(c, radius),
+Resource::Resource(Coordinate c, double radius, int quantity, double quantityEvolution) : NonSolid(c, radius),
     m_quantity(quantity), m_quantityEvolution(quantityEvolution)
 {
     m_maxQuantity = m_quantity;
